refactor(crypto): rename CryptoElement to BaseKey and simplify structure
- Renamed CryptoElement class to BaseKey for clarity and consistency.
- Updated constructor and method signatures to reflect the new class name.
- Removed unused fields and methods to streamline the code.

BREAKING CHANGE: The CryptoElement class has been renamed to BaseKey.
Impact: Existing references to CryptoElement will need to be updated to BaseKey.
Migration: Update all instances of CryptoElement in the codebase to BaseKey.
